requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
waitress==2.1.2
requests==2.31.0
PyJWT[crypto]==2.8.0
